requirements.txt
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
tzlocal==1.5.1
virtualenv==16.4.3
